+++ d2/src/main/java/com/linkedin/d2/balancer/simple/ClusterInfoItem.java
@@ -19,39 +19,54 @@ package com.linkedin.d2.balancer.simple;
 import com.linkedin.d2.balancer.LoadBalancerStateItem;
 import com.linkedin.d2.balancer.properties.ClusterProperties;
 import com.linkedin.d2.balancer.properties.FailoutProperties;
+import com.linkedin.d2.balancer.util.canary.CanaryDistributionProvider;
 import com.linkedin.d2.balancer.util.partitions.PartitionAccessor;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 
 /**
  * We put together the cluster properties and the partition accessor for a cluster so that we don't have to
  * maintain two separate maps (which have to be in sync all the time)
  */
-class ClusterInfoItem
+public class ClusterInfoItem
 {
   private final LoadBalancerStateItem<ClusterProperties> _clusterPropertiesItem;
   private final LoadBalancerStateItem<PartitionAccessor> _partitionAccessorItem;
   private final LoadBalancerStateItem<FailoutProperties> _failoutPropertiesItem;
 
-  ClusterInfoItem(SimpleLoadBalancerState simpleLoadBalancerState, ClusterProperties clusterProperties, PartitionAccessor partitionAccessor,
-                  FailoutProperties failoutProperties)
+  public ClusterInfoItem(SimpleLoadBalancerState simpleLoadBalancerState, ClusterProperties clusterProperties, PartitionAccessor partitionAccessor)
+  {
+    this(simpleLoadBalancerState, clusterProperties, partitionAccessor, CanaryDistributionProvider.Distribution.STABLE, null);
+  }
+
+  public ClusterInfoItem(
+      SimpleLoadBalancerState simpleLoadBalancerState,
+      ClusterProperties clusterProperties,
+      PartitionAccessor partitionAccessor,
+      @Nonnull
+      CanaryDistributionProvider.Distribution distribution,
+      @Nullable FailoutProperties failoutProperties)
   {
     long version = simpleLoadBalancerState.getVersionAccess().incrementAndGet();
     _clusterPropertiesItem = new LoadBalancerStateItem<>(clusterProperties,
-      version,
-      System.currentTimeMillis());
+        version,
+        System.currentTimeMillis(),
+        distribution);
     _partitionAccessorItem = new LoadBalancerStateItem<>(partitionAccessor,
-      version,
-      System.currentTimeMillis());
+        version,
+        System.currentTimeMillis());
     _failoutPropertiesItem = new LoadBalancerStateItem<>(failoutProperties,
       version,
       System.currentTimeMillis());
   }
 
-  LoadBalancerStateItem<ClusterProperties> getClusterPropertiesItem()
+
+  public LoadBalancerStateItem<ClusterProperties> getClusterPropertiesItem()
   {
     return _clusterPropertiesItem;
   }
 
-  LoadBalancerStateItem<PartitionAccessor> getPartitionAccessorItem()
+  public LoadBalancerStateItem<PartitionAccessor> getPartitionAccessorItem()
   {
     return _partitionAccessorItem;
   }